You can now add a view from a UINavigationController as a page to HGPageScroller.
In addition, you can use your UINavigationController's navigationBar as a custom header view for HGPageScroller, so that when the page is selected, the navigation view will become visible, and you can navigate forwards/backwards inside it. You can then deselect the page at any point to go back to deck mode. The last view you were in will replace the old one in the page scroller.